Performance updates that don't restructure the code (#3)
Just a couple of fixes to improve the performance.

The `self._ip_address_cache` was not actually being used.  
- The `get_reverse_dns` call was not passing on the cache parameter that
`get_ip_address_info` receives.
- The cache was being considered to be undefined because it was empty,
so entries were not being added to make it not empty.

The `dns_timeout` was not being passed all of the way from the options,
resulting in 6.0 seconds always being used for mail messages. (I
consider setting the timeout to 2.0 seconds to be a performance
improvement.)

I also made the update to the `Dockerfile` to work with the new source
folder structure.